cemese/requirements.txt
Django==1.3
FeinCMS==1.4.2
django-photologue==2.3
django-taggit==0.9.3
django-extensions==0.6
django-ckeditor==0.0.9
South==0.7.3
django-mptt==0.4.2

